refactor(link): move some links logic into its own file
